server/requirements.txt
# Flashcard App Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication
passlib>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
itsdangerous>=2.1.0
email-validator>=2.1.0

# OAuth
authlib>=1.3.0
httpx>=0.27.0

# Text-to-Speech
edge-tts>=6.1.0
